// ui/src/generated/types/user.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * DeviceHub
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 1.3.10
 */
import type { UserAdbKeysItem } from './userAdbKeysItem'
import type { UserGroupDeviceData } from './userGroupDeviceData'
import type { UserGroups } from './userGroups'
import type { UserSettings } from './userSettings'

export interface User {
  _id?: string
  acceptedPolicy?: boolean
  adbKeys?: UserAdbKeysItem[]
  createdAt?: string
  email?: string
  forwards?: unknown[]
  group?: string
  groupDeviceData?: UserGroupDeviceData
  groups?: UserGroups
  ip?: string
  lastLoggedInAt?: string
  name?: string
  privilege?: string
  settings?: UserSettings
}